% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{edge_table}
\alias{edge_table}
\title{edge_table}
\usage{
edge_table(tweet_df, text, screenName, ...)
}
\arguments{
\item{tweet_df}{Required. Data frame of tweets}

\item{text}{Required. Column name of tweets within tweets_df}

\item{screenName}{Required. User name or ID column name of users.}

\item{...}{Any other columns to be passed on to the edge_table, otpional.}
}
\value{
Returns a table of edges; source, target, loop, ...
}
\description{
Builds a table of edges (source, target) from a list of tweets by subsetting @tags from the text.
}
\details{
The edges function takes in a data frame of tweets typically obtained via the twitter API, scrapes the content of tweets to subset the @tags and form a table of edges. @tags are subsets of regular expressions starting with at-signs (@) until first space (" ").
Note that the table of edges returned is meant for a directed graph.
If no @tags are mentioned in the tweets (text) then the user name (screenName) is repeated to form a self-loop.
Self-loops can be easily filtered using the boolean loop variable (see return). Failed @tags i.e.:@ tag are not caught and will instead produce a self-loop.
}
\examples{
\dontrun{

names(tw_table)
[1] "text"          "favorited"     "favoriteCount" "replyToSN"     "created"
[6] "truncated"     "replyToSID"    "id"            "replyToUID"    "statusSource"
[11] "screenName"    "retweetCount"  "isRetweet"     "retweeted"     "longitude"
[16] "latitude"

edges_table <- edge_table(tweet_df=tw_table, text="text", screenName="screenName", "longitude", "latitude")

names(edges_table)
[1] "source"    "target"    "longitude" "latitude"

}
}
\seealso{
\href{http://cran.r-project.org/web/packages/twitteR/twitteR.pdf}{twitteR} and \href{http://cran.r-project.org/web/packages/streamR/streamR.pdf}{streamR} packages wherefrom the data (tweets_df) can be obtained.
}

