% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{gt_edges}
\alias{gt_edges}
\alias{gt_edges_}
\title{Edges}
\usage{
gt_edges(data, tweets, source, id, ...)

gt_edges_(data, tweets = "text", source = "screen_name", id = "status_id",
  ...)
}
\arguments{
\item{data}{Data.frame of tweets, usually returned by the \code{rtweet} package.}

\item{tweets}{Column containing tweets.}

\item{source}{Author of tweets.}
}
\description{
Get edges from data.frame of tweets.
}
\examples{
# simulate dataset
tweets <- tibble(
  text = c("I tweet @you about @him", 
           "I tweet @me about @you"),
  screen_name = c("me", "him")
)

tweets \%>\% 
  gt_edges(text, screen_name)

}
