% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhaul.R
\name{gt_edges_}
\alias{gt_edges_}
\alias{gt_edges_hashes}
\alias{gt_edges_hashes_}
\title{Deprecated Functions}
\usage{
gt_edges_(data, tweets = "text", source = "screen_name",
  id = "status_id", ...)

gt_edges_hashes(data, hashtags, tl = TRUE)

gt_edges_hashes_(data, hashtags = "hashtags", tl = TRUE)
}
\arguments{
\item{data}{Data.frame of tweets, usually returned by the \code{rtweet} package.}

\item{tweets}{Column containing tweets.}

\item{source}{Author of tweets.}

\item{id}{Unique id.}

\item{...}{any other column name, see examples.}

\item{hashtags}{Column containing co-mentions.}

\item{tl}{Set to \code{TRUE} to convert hashtags to lower case.}
}
\description{
These functions are deprecated, see \code{\link{gt_edges}} and \code{\link{gt_co_edges}}.
}
