% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhaul.R
\name{gt_collect}
\alias{gt_collect}
\title{Collect}
\usage{
gt_collect(gt)
}
\arguments{
\item{gt}{An object of class \code{graphTweets} as returned by \code{\link{gt_edges}}.}
}
\value{
A list of 1) edges and 2) nodes as \link[dlyr]{tibble}.
}
\description{
Collect
}
