% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhaul.R
\name{gt_graph}
\alias{gt_graph}
\title{Graph}
\usage{
gt_graph(gt)
}
\arguments{
\item{gt}{An object of class \code{graphTweets} as returned by \code{\link{gt_edges}}.}
}
\value{
An object of class \code{igraph}.
}
\description{
Graph
}
\examples{
# simulate dataset
tweets <- data.frame(
  text = c("I tweet @you about @him", 
           "I tweet @me about @you"),
  screen_name = c("me", "him"),
  retweet_count = c(19, 5),
  status_id = c(1, 2),
  stringsAsFactors = FALSE
)

tweets \%>\% 
  gt_edges(text, screen_name, status_id) \%>\% 
  gt_nodes() \%>\% 
  gt_graph() -> net

}
