% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R, R/overhaul.R
\name{gt_nodes}
\alias{gt_nodes}
\alias{gt_nodes}
\title{Nodes}
\usage{
gt_nodes(gt, meta = FALSE)

gt_nodes(gt, meta = FALSE)
}
\arguments{
\item{gt}{An object of class \code{graphTweets} as returned by \code{\link{gt_edges}}.}
}
\description{
Nodes

Nodes
}
