% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{node_table}
\alias{node_table}
\title{node_table}
\usage{
node_table(edge_table, ...)
}
\arguments{
\item{edge_table}{Required - data.frame. First column is source node, second is target node.}

\item{...}{Any other column names (meta-data)}
}
\value{
Returns table of nodes/vertices; first column is nodes' name or ID following columns are optional args. Meant to be used as meta-data for graph.
}
\description{
Builds table of nodes from an edge table (data.frame). Also known as vertices. Ideal to add metadata to igraph objects/graphml files.
}
\details{
Looks for uniques in both source and target columns of the edge table it is fed. Note that other (optional) args only apply to the source, not the target; will return NA. Assumes the first column of the edge_table are the sources of edges and the second are the targets.
}
\examples{
\dontrun{

names(edge_table)
[1] "source"    "target"    "longitude" "latitude"  "self_loop"

node_table <- node_table(edge_table, "longitude", "latitude")

names(node_table)
[1] "screenName" "longitude"  "latitude"
}
}
\seealso{
\code{\link{edge_table}}
}

