% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GraphTweets.R
\name{edge_table}
\alias{edge_table}
\title{edge_table}
\usage{
edge_table(tweet_df, text, screenName, strLength = FALSE, ...)
}
\arguments{
\item{tweet_df}{Data frame of tweets which includes text and screenNames, required.}

\item{text}{Column name of tweets within tweet_df, must be of character class, required.}

\item{screenName}{User name or ID column, must be of character class, required.}

\item{strLength}{Defaults to FALSE. Shorten length of @tags (see details), maximum number of characters, optional.}

\item{...}{Any other columns to be passed on to the edge_table function, optional.}
}
\value{
Returns a table of edges; source, target, loop, ...
}
\description{
Builds a table of edges (source, target) from a list of tweets by subsetting @tags from the text.
}
\details{
The edges function takes in a data frame of tweets -typically obtained from the twitter Search or Streaming API-, scrapes the content of tweets to subset the @tags subsequently forming a table of edges. @tags are subsets of regular expressions between at-signs (@) and first space (" ").
Note that the table of edges returned is meant for a directed graph.
If no @tags are mentioned in the tweets (text) then the user name or ID (screenName) is repeated to form a self-loop. Self-loop can be identified by loop variables (boolean). Failed @tags i.e.:@ tag are not caught and will instead produce a self-loop.
Node labels can be shortened using the strLength parameters. This is useful for non-latin alphabet where nodes may be wrongly identified.
}
\examples{
\dontrun{

names(tw_table)
[1] "text"          "favorited"     "favoriteCount" "replyToSN"     "created"
[6] "truncated"     "replyToSID"    "id"            "replyToUID"    "statusSource"
[11] "screenName"    "retweetCount"  "isRetweet"     "retweeted"     "longitude"
[16] "latitude"

edges_table <- edge_table(tweet_df = tw_table, text = "text", screenName = "screenName", "retweetCount", strLength = FALSE)

# output
names(edges_table)
[1] "source"    "target"   "loop"    "longitude"    "latitude"

}
}
\seealso{
\href{http://cran.r-project.org/web/packages/twitteR/twitteR.pdf}{twitteR} and \href{http://cran.r-project.org/web/packages/streamR/streamR.pdf}{streamR} packages wherefrom the data (tweets_df) can be obtained.
}

