% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhaul.R
\name{gt_edges}
\alias{gt_edges}
\alias{gt_edges_bind}
\alias{gt_co_edges}
\alias{gt_co_edges_bind}
\title{Edges}
\usage{
gt_edges(data, source, target, ..., tl = TRUE)

gt_edges_bind(gt, source, target, ..., tl = TRUE)

gt_co_edges(data, col, tl = TRUE)

gt_co_edges_bind(gt, col, tl = TRUE)
}
\arguments{
\item{data}{Data.frame of tweets, usually returned by the \code{rtweet} package.}

\item{source}{Author of tweets.}

\item{target}{Edges target.}

\item{...}{any other column name, see examples.}

\item{tl}{Set to \code{TRUE} to convert hashtags to lower case.}

\item{gt}{An object of class \code{graphTweets} as returned by \code{\link{gt_edges}} and \code{\link{gt_nodes}}.}

\item{col}{Column containing co-mentions.}
}
\description{
Get edges from data.frame of tweets.
}
