% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhaul.R
\name{gt_collect}
\alias{gt_collect}
\title{Collect}
\usage{
gt_collect(gt)
}
\arguments{
\item{gt}{An object of class \code{graphTweets} as returned by \code{\link{gt_edges}}.}
}
\value{
A named list of \link[dplyr]{tibble} 1) edges and 2) nodes.
}
\description{
Collect
}
\examples{
# simulate dataset
tweets <- data.frame(
  text = c("I tweet @you about @him", 
           "I tweet @me about @you"),
  screen_name = c("me", "him"),
  retweet_count = c(19, 5),
  status_id = c(1, 2),
  stringsAsFactors = FALSE
)

tweets \%>\% 
  gt_edges(text, screen_name, status_id) \%>\% 
  gt_nodes() \%>\% 
  gt_collect() -> net

}
