% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhaul.R
\name{gt_nodes}
\alias{gt_nodes}
\title{Nodes}
\usage{
gt_nodes(gt, meta = FALSE)
}
\arguments{
\item{gt}{An object of class \code{graphTweets} as returned by \code{\link{gt_edges}} and \code{\link{gt_nodes}}.}

\item{meta}{Set to \code{TRUE} to add meta data to nodes.}
}
\value{
An object of class \code{graphTweets}.
}
\description{
Get nodes from a \code{graphTweets} object.
}
