% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GraphTweets.R
\name{node_table}
\alias{node_table}
\title{node_table}
\usage{
node_table(edge_table, ...)
}
\arguments{
\item{edge_table}{data.frame. Assumes first column is source node, second is target node, required.}

\item{...}{Any other column names (meta-data)}
}
\value{
Returns table of nodes/vertices; first column is nodes' name or ID following columns are optional args. Meant to be used as meta-data for graph.
}
\description{
Builds table of nodes|vertices from an edge table (data.frame). Use to metadata to igraph objects/graphml files.
}
\details{
Looks for uniques in both source and target columns of the edge table it is fed. Note that other args (optional) only apply to the source node; will return NA for target nodes. Assumes first column of the edge_table are source nodes and the second are target nodes.
}
\examples{
\dontrun{

names(edge_table)
[1] "source"    "target"    "longitude" "latitude"  "self_loop"

node_table <- node_table(edge_table, "longitude", "latitude")

# output
names(node_table)
[1] "screenName" "longitude"  "latitude"
}
}
\seealso{
\code{\link{edge_table}}
}

